+++ src/main/java/com/finns/follow/mapper/FollowMapper.java
@@ -27,14 +27,14 @@ public interface FollowMapper {
      * @param user_no 사용자 번호
      * @return 팔로잉 사용자 번호 목록
      */
-    List<MemberVO> selectFollowingByUserNo(int user_no);
+    List<MemberVO> selectFollowingByUserNo(long user_no);
 
     /**
      * 특정 사용자를 팔로우하는 사용자 목록을 가져온다.
      * @param user_no 사용자 번호
      * @return 팔로워 사용자 번호 목록
      */
-    List<MemberVO> selectFollowerByUserNo(int user_no);
+    List<MemberVO> selectFollowerByUserNo(long user_no);
 
     /**
      * 팔로우 관계가 이미 존재하는지 확인한다.
@@ -43,6 +43,6 @@ public interface FollowMapper {
      */
     int checkFollowExists(@Param("user_no") long user_no, @Param("to_user_no") long to_user_no);
 
-    int countFollowers(int user_no);
-    int countFollowing(int user_no);
+    int countFollowers(long user_no);
+    int countFollowing(long user_no);
 }
